Move STLC normalization to own module
